Move RuleIndex to ncore
